+++ docs_src/databases/edgy/middleware/example3.py
@@ -16,14 +16,10 @@ class AppAuthMiddleware(JWTAuthMiddleware):
     Overriding the JWTAuthMiddleware
     """
 
-    jwt_config = JWTConfig(
-        signing_key=settings.secret_key, auth_header_types=["Bearer", "Token"]
-    )
+    jwt_config = JWTConfig(signing_key=settings.secret_key, auth_header_types=["Bearer", "Token"])
 
     def __init__(self, app: "ASGIApp"):
-        super().__init__(
-            app, config=self.jwt_config, user_model=load("myapp.models.User")
-        )
+        super().__init__(app, config=self.jwt_config, user_model=load("myapp.models.User"))
 
 
 class AppSettings(EsmeraldAPISettings):