+++ docs_src/databases/saffier/middleware/example1.py
@@ -5,9 +5,7 @@ from esmerald.contrib.auth.saffier.middleware import JWTAuthMiddleware
 from monkay import load
 from lilya.middleware import DefineMiddleware as LilyaMiddleware
 
-jwt_config = JWTConfig(
-    signing_key=settings.secret_key, auth_header_types=["Bearer", "Token"]
-)
+jwt_config = JWTConfig(signing_key=settings.secret_key, auth_header_types=["Bearer", "Token"])
 
 jwt_auth_middleware = LilyaMiddleware(
     JWTAuthMiddleware,