+++ esmerald/routing/router.py
@@ -283,9 +283,7 @@ class BaseRouter(LilyaRouter):
             ),
         ] = None,
         routes: Annotated[
-            Optional[
-                Sequence[Union[APIGateHandler, Include, HTTPHandler, WebSocketHandler]]
-            ],
+            Optional[Sequence[Union[APIGateHandler, Include, HTTPHandler, WebSocketHandler]]],
             Doc(
                 """
                 A `list` of esmerald routes. Those routes may vary and those can
@@ -681,9 +679,7 @@ class BaseRouter(LilyaRouter):
         if "app" in scope:
             raise NotFound(status_code=status.HTTP_404_NOT_FOUND)
         else:
-            response = JSONResponse(
-                {"detail": "Not Found"}, status_code=status.HTTP_404_NOT_FOUND
-            )
+            response = JSONResponse({"detail": "Not Found"}, status_code=status.HTTP_404_NOT_FOUND)
         await response(scope, receive, send)
 
     def path_for(self, name: str, **path_params: Any) -> URLPath:
@@ -702,9 +698,7 @@ class BaseRouter(LilyaRouter):
 
         raise NoMatchFound(name, path_params)
 
-    def add_event_handler(
-        self, event_type: str, func: Callable
-    ) -> None:  # pragma: no cover
+    def add_event_handler(self, event_type: str, func: Callable) -> None:  # pragma: no cover
         assert event_type in ("startup", "shutdown")
 
         if event_type == "startup":
@@ -757,9 +751,7 @@ class BaseRouter(LilyaRouter):
 
         if isinstance(value, (Include, Gateway, WebSocketGateway, WebhookGateway)):
             if not value.parent and not override:
-                value.parent = cast(
-                    "Union[Router, Include, Gateway, WebSocketGateway]", self
-                )
+                value.parent = cast("Union[Router, Include, Gateway, WebSocketGateway]", self)
 
         if isinstance(value, (Gateway, WebSocketGateway, WebhookGateway)):
             if not is_class_and_subclass(value.handler, View) and not isinstance(
@@ -835,17 +827,15 @@ class Router(BaseRouter):
         if not value.handler.parent:  # pragma: no cover
             value.handler(parent=self)
 
-        route_handlers: List[Union[HTTPHandler, WebSocketHandler]] = (
-            value.handler.get_routes(
-                path=value.path,
-                middleware=value.middleware,
-                interceptors=value.interceptors,
-                permissions=value.permissions,
-                exception_handlers=value.exception_handlers,
-                include_in_schema=value.include_in_schema,
-                before_request=value.before_request,
-                after_request=value.after_request,
-            )
+        route_handlers: List[Union[HTTPHandler, WebSocketHandler]] = value.handler.get_routes(
+            path=value.path,
+            middleware=value.middleware,
+            interceptors=value.interceptors,
+            permissions=value.permissions,
+            exception_handlers=value.exception_handlers,
+            include_in_schema=value.include_in_schema,
+            before_request=value.before_request,
+            after_request=value.after_request,
         )
         if route_handlers:
             self.routes.extend(route_handlers)
@@ -1238,9 +1228,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -1352,9 +1340,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -1466,9 +1452,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -1580,9 +1564,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -1694,9 +1676,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -1808,9 +1788,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -1922,9 +1900,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -2036,9 +2012,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -2161,9 +2135,7 @@ class Router(BaseRouter):
                 after_request=after_request,
             )
             handler.fn = func
-            self.add_route(
-                path=path, handler=handler, name=name, interceptors=interceptors
-            )
+            self.add_route(path=path, handler=handler, name=name, interceptors=interceptors)
             return func
 
         return wrapper
@@ -2358,9 +2330,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         self._permissions: Union[List[Permission], VoidType] = Void
         self._dependencies: Dependencies = {}
 
-        self._response_handler: Union[
-            Callable[[Any], Awaitable[LilyaResponse]], VoidType
-        ] = Void
+        self._response_handler: Union[Callable[[Any], Awaitable[LilyaResponse]], VoidType] = Void
 
         self.parent: ParentType = None
         self.path = path
@@ -2393,9 +2363,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         self.dependencies: Dependencies = dependencies or {}
 
         if not description:
-            description = (
-                inspect.cleandoc(self.handler.__doc__ or "") if self.handler else ""
-            )
+            description = inspect.cleandoc(self.handler.__doc__ or "") if self.handler else ""
 
         self.description = description
 
@@ -2444,9 +2412,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
                 )
 
             if not is_status_code_allowed(status_code):
-                raise OpenAPIException(
-                    detail="The status is not a valid OpenAPI status response."
-                )
+                raise OpenAPIException(detail="The status is not a valid OpenAPI status response.")
 
     async def __call__(self, scope: Scope, receive: Receive, send: Send) -> Any:
         await self.handle_dispatch(scope=scope, receive=receive, send=send)
@@ -2511,9 +2477,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
                 filtered_cookies.append(cookie)
         return filtered_cookies
 
-    async def handle_dispatch(
-        self, scope: "Scope", receive: "Receive", send: "Send"
-    ) -> None:
+    async def handle_dispatch(self, scope: "Scope", receive: "Receive", send: "Send") -> None:
         """
         Handles the dispatching of a request.
         This method processes the incoming request by performing the following steps:
@@ -2631,8 +2595,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
             )
         if (
             self.status_code < status.HTTP_200_OK
-            or self.status_code
-            in {status.HTTP_204_NO_CONTENT, status.HTTP_304_NOT_MODIFIED}
+            or self.status_code in {status.HTTP_204_NO_CONTENT, status.HTTP_304_NOT_MODIFIED}
         ) and return_annotation not in [NoReturn, None]:
             raise ImproperlyConfigured(
                 "A status code 204, 304 or in the range below 200 does not support a response body."
@@ -2662,9 +2625,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
                     raise ImproperlyConfigured(
                         f"'{special}' argument unsupported when only body-less methods like 'GET' and 'HEAD' are handled"
                     )
-                elif not is_optional_union(
-                    self.handler_signature.parameters[special].annotation
-                ):
+                elif not is_optional_union(self.handler_signature.parameters[special].annotation):
                     raise ImproperlyConfigured(
                         f"'{special}' argument must be optional when body-less methods like 'GET' and 'HEAD' are handled"
                     )
@@ -2689,9 +2650,7 @@ class HTTPHandler(Dispatcher, OpenAPIFieldInfoMixin, LilyaPath):
         """
 
         if SOCKET in self.handler_signature.parameters:
-            raise ImproperlyConfigured(
-                "The 'socket' argument is not supported with http handlers"
-            )
+            raise ImproperlyConfigured("The 'socket' argument is not supported with http handlers")
 
     def validate_handler(self) -> None:
         self.check_handler_function()
@@ -2863,9 +2822,7 @@ class WebSocketHandler(Dispatcher, LilyaWebSocketPath):
         self._permissions: Union[List[Permission], VoidType] = Void
         self._lilya_permissions: Union[List[DefinePermission], VoidType] = Void
         self._dependencies: Dependencies = {}
-        self._response_handler: Union[
-            Callable[[Any], Awaitable[LilyaResponse]], VoidType
-        ] = Void
+        self._response_handler: Union[Callable[[Any], Awaitable[LilyaResponse]], VoidType] = Void
         self._interceptors: Union[List[Interceptor], VoidType] = Void
         self.interceptors: Sequence[Interceptor] = []
         self.handler = handler
@@ -2914,17 +2871,13 @@ class WebSocketHandler(Dispatcher, LilyaWebSocketPath):
         self.validate_reserved_words(signature=signature)
 
         if SOCKET not in signature.parameters:
-            raise ImproperlyConfigured(
-                "Websocket handlers must set a 'socket' argument."
-            )
+            raise ImproperlyConfigured("Websocket handlers must set a 'socket' argument.")
         if not is_async_callable(self.fn):
             raise ImproperlyConfigured(
                 "Functions decorated with 'asgi, get, patch, put, post and delete' must be async functions."
             )
 
-    async def handle_dispatch(
-        self, scope: "Scope", receive: "Receive", send: "Send"
-    ) -> None:
+    async def handle_dispatch(self, scope: "Scope", receive: "Receive", send: "Send") -> None:
         """
         Handles the dispatching of a request.
         This method processes the incoming request by handling interceptors,
@@ -2998,14 +2951,10 @@ class WebSocketHandler(Dispatcher, LilyaWebSocketPath):
         signature_model = get_signature(self)
         kwargs = await self.websocket_parameter_model.to_kwargs(connection=websocket)
         for dependency in self.websocket_parameter_model.dependencies:
-            kwargs[dependency.key] = (
-                await self.websocket_parameter_model.get_dependencies(
-                    dependency=dependency, connection=websocket, **kwargs
-                )
+            kwargs[dependency.key] = await self.websocket_parameter_model.get_dependencies(
+                dependency=dependency, connection=websocket, **kwargs
             )
-        return await signature_model.parse_values_for_connection(
-            connection=websocket, **kwargs
-        )
+        return await signature_model.parse_values_for_connection(connection=websocket, **kwargs)
 
 
 class Include(LilyaInclude):
@@ -3128,9 +3077,7 @@ class Include(LilyaInclude):
             ),
         ] = None,
         routes: Annotated[
-            Optional[
-                Sequence[Union[APIGateHandler, Include, HTTPHandler, WebSocketHandler]]
-            ],
+            Optional[Sequence[Union[APIGateHandler, Include, HTTPHandler, WebSocketHandler]]],
             Doc(
                 """
                 A global `list` of esmerald routes. Those routes may vary and those can
@@ -3386,14 +3333,10 @@ class Include(LilyaInclude):
             raise ImproperlyConfigured("It can only be namespace or routes, not both.")
 
         if namespace and not isinstance(namespace, str):
-            raise ImproperlyConfigured(
-                "Namespace must be a string. Example: 'myapp.routes'."
-            )
+            raise ImproperlyConfigured("Namespace must be a string. Example: 'myapp.routes'.")
 
         if pattern and not isinstance(pattern, str):
-            raise ImproperlyConfigured(
-                "Pattern must be a string. Example: 'route_patterns'."
-            )
+            raise ImproperlyConfigured("Pattern must be a string. Example: 'route_patterns'.")
 
         if pattern and routes:
             raise ImproperlyConfigured("Pattern must be used only with namespace.")
@@ -3534,9 +3477,7 @@ class Include(LilyaInclude):
                 route = WebSocketGateway(handler=route)
 
             if not isinstance(route, (Include, Gateway, WebSocketGateway)):
-                raise ImproperlyConfigured(
-                    "The route must be of type Gateway or Include"
-                )
+                raise ImproperlyConfigured("The route must be of type Gateway or Include")
 
             route.parent = self
             if isinstance(route, Include):
@@ -3548,9 +3489,7 @@ class Include(LilyaInclude):
                 routing.append(route)
                 continue
 
-            if is_class_and_subclass(route.handler, View) or isinstance(
-                route.handler, View
-            ):
+            if is_class_and_subclass(route.handler, View) or isinstance(route.handler, View):
                 if not route.handler.parent:
                     route.handler = route.handler(parent=self)